Move preApprovals tests
